requirements-test.txt
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-playwright==0.4.0